@@ -7,7 +7,7 @@
     "clean": "tsc --build --clean",
     "build": "tsc",
     "watch": "tsc -w",
-    "test": "jest",
+    "test": "jest --passWithNoTests --updateSnapshot",
     "cdk": "cdk",
     "doc": "typedoc --sort source-order --readme README.md",
     "format": "eslint '**/*.{js,ts,json}' --quiet --fix",
@@ -15,8 +15,8 @@
     "postclean": "rm -rf ./dist"
   },
   "dependencies": {
-    "aws-cdk-lib": "2.176.0",
-    "aws4embeddedlinux-cdk-lib": "adadouche/aws4embeddedlinux-ci#dev-adadouche",
+    "aws-cdk-lib": "^2.186.0",
+    "aws4embeddedlinux-cdk-lib": "github:aws4embeddedlinux/aws4embeddedlinux-ci",
     "constructs": "^10.4.2",
     "source-map-support": "^0.5.21"
   },
@@ -28,11 +28,12 @@
     "@types/node": "20.14.8",
     "@typescript-eslint/eslint-plugin": "^8.18.0",
     "@typescript-eslint/parser": "^8.18.0",
-    "aws-cdk": "2.176.0",
+    "aws-cdk": "^2.1006.0",
     "cdk-nag": "^2.35.0",
     "eslint": "^9.16.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-simple-import-sort": "^12.1.1",
     "eslint-plugin-tsdoc": "^0.4.0",
     "globals": "^15.13.0",
     "jest": "^29.7.0",
@@ -40,7 +41,7 @@
     "ts-jest": "^29.2.5",
     "ts-node": "^10.9.2",
     "typedoc": "^0.27.4",
-    "typescript": "~5.7.2",
+    "typescript": "~5.8.2",
     "typescript-eslint": "^8.18.0"
   },
   "resolutions": {
@@ -49,5 +50,5 @@
   "overrides": {
     "glob": "^9.3.5"
   },
-  "packageManager": "yarn@4.6.0"
+  "packageManager": "yarn@4.7.0"
 }
